+++ core/test/org/ice1000/jimgui/features/DateChooserTest.java
@@ -12,7 +12,7 @@ public class DateChooserTest {
     try (JImGui imGui = new JImGui(); NativeTime time = new NativeTime()) {
       while (!imGui.windowShouldClose()) {
         imGui.initNewFrame();
-        imGui.text("Picked text: " + time.toAbsoluteSeconds() * 1000);
+        imGui.text("Picked text: " + time.accessAbsoluteSeconds() * 1000);
         imGui.text("Current time text: " + System.currentTimeMillis());
         if (imGui.dateChooser("Pick a time", time)) imGui.text("Text");
         imGui.render();